Music
Burning Candle sounds: should be on loop and with 3d spatial sound being played at the pivot of the candle. this means that the player would hear or not the sound depeding on the distance to the candle.

Monster Chant: this sound should only play while the monster does not chase the player

Monster Stomping: monster footsteps, play regarles if chasing or not

Main Theme: background sound to be played in the main menu

Atmosphere 02: should be played as background music while playing.

Heavy Breathing: only when the player runs

Running: only when the player runs.

Footsteps: only when the player walks